Adds button to open developer options.
Don't keep activities now only shows the value on the setting since it wasn't really changing the setting (although setting seemed to change, the behaviour was maintained).
